pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surferbro"
version = "0.1.0"
description = "Reinforcement Learning Surfing Simulator with Advanced Ocean Physics"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SurferBro Team"}
]
keywords = ["reinforcement-learning", "surfing", "physics-simulation", "gymnasium", "ocean-waves"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "gymnasium>=0.28.0",
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
    "pybullet>=3.2.0",
    "pygame>=2.5.0",
    "flask>=2.3.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "tensorboard>=2.13.0",
    "trimesh>=3.21.0",
    "pillow>=9.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
]

[project.scripts]
surferbro-train = "surferbro.training.train:main"
surferbro-evaluate = "surferbro.training.evaluate:main"
surferbro-oceanscope = "surferbro.oceanscope.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["surferbro*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
